add update to product details, add pattern and error message to headerDetails components, fix render headerDetails in LinedetailsPage(an issue related to changing headerDetailsComponent)

+++ src/templates/LineDetailsPage/LineDetailsPage.js
@@ -15,13 +15,15 @@ export const LineDetailsPage = props => {
     const [line, setLine] = useState('');
     const [message, setMessage] = useState({ isOpen: false, text: [] });
     const [isExistingId, setIsExistingId] = useState(true);
-
+    const [isSubmiting, setIsSubmiting] = useState(false);
     const content = [
         {
             label: 'Linia',
             name: 'name',
             value: line.name,
             edit: true,
+            pattern: '.{4,30}',
+            errorMessage: 'Nazwa linii musi zawierać od 4 do 30 znaków'
         },
         {
             label: 'Numer linii',
@@ -36,26 +38,43 @@ export const LineDetailsPage = props => {
     };
 
     const handleUpdateLine = (data) => {
+        setIsSubmiting(true);
         LineService.update(params.idLine, data)
             .then(res => {
+                setIsSubmiting(false);
                 setLine(res);
                 setMessage({ isOpen: true, text: ['Zaktualizowano'] });
             })
-            .catch(err => setMessage({ isOpen: true, text: ['Wystąpił błąd łączności', `Błąd ${err}`, 'spróbuj ponownie'] }));
+            .catch(err => {
+                setIsSubmiting(false);
+                setMessage({ isOpen: true, text: ['Wystąpił błąd łączności', `Błąd ${err}`, 'spróbuj ponownie'] });
+            });
     }
     const handleRemoveLine = () => {
+        const answer = window.confirm("Czy na pewno chcesz usunąć produkt?");
+        if (!answer) return
+        setIsSubmiting(true);
         LineService.remove(params.idLine)
             .then(res => {
                 setMessage({ isOpen: true, text: ['Usunieto linię'] });
-                setTimeout(() => setIsExistingId(false), 2000)
+                setTimeout(() => setIsExistingId(false), 2000);
+                setIsSubmiting(false);
             })
-            .catch(err => setMessage({ isOpen: true, text: ['Wystąpił błąd łączności', `Błąd ${err}`, 'spróbuj ponownie'] }))
+            .catch(err => {
+                setIsSubmiting(false);
+                setMessage({ isOpen: true, text: ['Wystąpił błąd łączności', `Błąd ${err}`, 'spróbuj ponownie'] })
+            });
     }
 
     useEffect(() => {
+        setIsSubmiting(true);
         LineService.get(params.idLine)
-            .then(data => setLine(data))
+            .then(data => {
+                setIsSubmiting(false);
+                setLine(data);
+            })
             .catch(status => {
+                setIsSubmiting(false);
                 if (status === 404) {
                     setMessage({ isOpen: true, text: ['Sprawdź poprawność linku', 'Linia nie istnieje'] })
                     setTimeout(() => setIsExistingId(false), 3000)
@@ -70,7 +89,7 @@ export const LineDetailsPage = props => {
             < section className={props.className} >
                 <DialogMessage open={message.isOpen} close={handleCloseMessage} messages={message.text} />
                 <Card container component={Grid}>
-                    <HeaderDetails content={content} onChange={handleUpdateLine} onRemove={handleRemoveLine} />
+                    <HeaderDetails content={content} onChange={handleUpdateLine} onRemove={handleRemoveLine} isSubmiting={isSubmiting} />
                     <HeaderDetailsCircles title='Produkcja' id={params.idLine} type='LINE' setMessage={setMessage} />
                 </Card>
                 <LineTable id={params.idLine} type='LINE' />
